проблема в типах что там char а надо u32 но обрати внимание что я видел функцию в print что так же char * но все прекрано работало. все говно сейчас в Console Syntax

+++ src/Includes/libs/datastruct.h
@@ -2,10 +2,10 @@
 #include <types.h>
 
 typedef struct {
-  int (*put)(u8 character);
-  int (*get)(u8 *linkchar);
-  u8 buffer[8192];
+  int (*put)(u32 character);
+  int (*get)(u32 *linkchar);
+  u32 buffer[8192];
 } RoundBufferObgect;
 
-int PutIntoBuffer(u8 character);
-int GetgFromBuffer(u8 *linkchar);
+int PutIntoBuffer(u32 character);
+int GetgFromBuffer(u32 *linkchar);
